+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add conversation history size setting
+
 ## 1.0.0 - 2026-03-20
 ### Added
 - Add memory support to record conversations history between bot and players